jira/demo/requirements.txt
ConcurrentLogHandler==0.9.1
Django==2.2.6
django-cors-headers==3.1.1
django-debug-toolbar==2.1
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
psycopg2==2.8.4
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.0
